# Perf/remove noop casts