While removing `isinstance` calls, @rayanht added some noop casts to may mypy happy.

In my testing, the class check + the noop cast to satisfy mypy is slower than an isinstance call. 
In some cases, mypy can be satisfied with an assert, which means the optimized can run without the cast.

Very minor speedup, but mainly neater without the noop casts.

This branch
```
codegen         mean runtime:  139.25ms, runs:   164.62,  171.97,  133.90,  129.46,  135.23,  130.50,  128.49,  131.80,  132.25,  134.24
methodcache     mean runtime:  134.42ms, runs:   172.48,  124.85,  126.02,  124.91,  122.18,  122.96,  127.17,  130.37,  124.42,  168.88
```

Master
```
codegen         mean runtime:  147.10ms, runs:   168.93,  194.08,  144.76,  141.64,  132.99,  144.88,  130.67,  143.57,  135.97,  133.52
methodcache     mean runtime:  141.62ms, runs:   168.17,  125.07,  127.03,  128.15,  126.01,  122.48,  124.39,  128.67,  155.49,  210.71

```