@geohot 2023-06-29
Some of these improve the code quality, some make it worse. I'd only merge this if it's a clear code quality win