@roelofvandijk 2023-06-30
Updated.

I think this is worth it, see this example form the very busy tensor init. Note the cast in line 55.

master
```
Total time: 0.735203 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/tensor.py
Function: __init__ at line 41

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    41                                             @profile
    42                                             def __init__(self, data:Union[int, float, list, tuple, LazyBuffer, np.ndarray], device:Optional[str]=None, dtype:Optional[DType]=None, requires_grad:Optional[bool]=None):
    43    188815      49434.0      0.3      6.7      assert dtype is None or isinstance(dtype, DType), f"invalid dtype {dtype}"
    44    188815     115312.8      0.6     15.7      device = Device.canonicalize(device)
    45                                               # tensors have gradients, buffers do not
    46    188815      41352.6      0.2      5.6      self.grad: Optional[Tensor] = None
    47                                           
    48                                               # NOTE: this can be in three states. False and None: no gradient, True: gradient
    49                                               # None (the default) will be updated to True if it's put in an optimizer
    50    188815      33112.4      0.2      4.5      self.requires_grad: Optional[bool] = requires_grad
    51                                           
    52                                               # internal variables used for autograd graph construction
    53    188815      32631.8      0.2      4.4      self._ctx: Optional[Function] = None
    54    178486      47766.2      0.3      6.5      if data.__class__ is LazyBuffer:
    55    178486     131869.0      0.7     17.9        data = cast(LazyBuffer, data) # NOTE: this is a noop, it makes mypy happy
    56    178486      28598.4      0.2      3.9        assert dtype is None or dtype == data.dtype, "dtype doesn't match, and casting isn't supported"
    57    178486      67920.6      0.4      9.2        self.lazydata = data if data.device == device else LazyBuffer.loadop(LoadOps.FROM, data.shape, data.dtype, device, src=data)
    58    178486      27692.3      0.2      3.8        return
    59                                           
    60     10298       7829.1      0.8      1.1      if isinstance(data, (int, float)):
    61     10298     146372.1     14.2     19.9        self.lazydata = LazyBuffer.loadop(LoadOps.CONST, tuple(), dtype or Tensor.default_type, device, data)
    62     10298       2201.2      0.2      0.3        return
    63                                           
    64        30          9.0      0.3      0.0      if data.__class__ is list:
    65        30        440.8     14.7      0.1        data = np.array(data, dtype=(dtype or Tensor.default_type).np)
    66                                           
    67        31         42.9      1.4      0.0      if data.__class__ is np.ndarray:
    68        31         32.8      1.1      0.0        data = cast(np.ndarray, data)
    69        31       1749.2     56.4      0.2        data = LazyBuffer.fromCPU(data)
    70        31        829.0     26.7      0.1        self.lazydata = data if data.device == device else LazyBuffer.loadop(LoadOps.FROM, data.shape, data.dtype, device, src=data)
    71        31          7.0      0.2      0.0        return
    72                                           
    73                                               raise RuntimeError(f"can't create Tensor from {data}")
```

This branch
```
Total time: 0.605383 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/tensor.py
Function: __init__ at line 41

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    41                                             @profile
    42                                             def __init__(self, data:Union[int, float, list, tuple, LazyBuffer, np.ndarray], device:Optional[str]=None, dtype:Optional[DType]=None, requires_grad:Optional[bool]=None):
    43    188815      50771.7      0.3      8.4      assert dtype is None or isinstance(dtype, DType), f"invalid dtype {dtype}"
    44    188815     109645.7      0.6     18.1      device = Device.canonicalize(device)
    45                                               # tensors have gradients, buffers do not
    46    188815      51856.6      0.3      8.6      self.grad: Optional[Tensor] = None
    47                                           
    48                                               # NOTE: this can be in three states. False and None: no gradient, True: gradient
    49                                               # None (the default) will be updated to True if it's put in an optimizer
    50    188815      30687.0      0.2      5.1      self.requires_grad: Optional[bool] = requires_grad
    51                                           
    52                                               # internal variables used for autograd graph construction
    53    188815      32578.5      0.2      5.4      self._ctx: Optional[Function] = None
    54    178486      53988.0      0.3      8.9      if type(data) is LazyBuffer:
    55    178486      27955.6      0.2      4.6        assert dtype is None or dtype == data.dtype, "dtype doesn't match, and casting isn't supported"
    56    178486      64601.8      0.4     10.7        self.lazydata = data if data.device == device else LazyBuffer.loadop(LoadOps.FROM, data.shape, data.dtype, device, src=data)
    57    178486      30845.8      0.2      5.1        return
    58                                           
    59     10298       5426.6      0.5      0.9      if type(data) in (int, float):
    60     10298     141989.2     13.8     23.5        self.lazydata = LazyBuffer.loadop(LoadOps.CONST, tuple(), dtype or Tensor.default_type, device, data)
    61     10298       1868.5      0.2      0.3        return
    62                                           
    63        30          9.4      0.3      0.0      if data.__class__ is list:
    64        30        430.1     14.3      0.1        data = np.array(data, dtype=(dtype or Tensor.default_type).np)
    65                                           
    66        31         22.5      0.7      0.0      if type(data) is np.ndarray:
    67        31       1819.5     58.7      0.3        data = LazyBuffer.fromCPU(data)
    68        31        879.9     28.4      0.1        self.lazydata = data if data.device == device else LazyBuffer.loadop(LoadOps.FROM, data.shape, data.dtype, device, src=data)
    69        31          6.1      0.2      0.0        return
```